pyproject.toml
[tool.poetry]
name = "funcnodes-images"
version = "0.1.0"
description = ""
authors = ["Julian Kimmig <julian.kimmig@gmx.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
funcnodes = "^0.1.30"
numpy = "^1.26.4"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
opencv-python-headless = "^4.9.0.80"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
